modify services according spring data

+++ services/src/main/java/com/netcracker/komarov/services/impl/CardServiceImpl.java
@@ -1,8 +1,10 @@
 package com.netcracker.komarov.services.impl;
 
+import com.netcracker.komarov.dao.entity.Account;
 import com.netcracker.komarov.dao.entity.Card;
 import com.netcracker.komarov.dao.entity.UnlockCardRequest;
 import com.netcracker.komarov.dao.factory.RepositoryFactory;
+import com.netcracker.komarov.dao.repository.AccountRepository;
 import com.netcracker.komarov.dao.repository.CardRepository;
 import com.netcracker.komarov.dao.repository.UnlockCardRequestRepository;
 import com.netcracker.komarov.services.interfaces.CardService;
@@ -14,34 +16,54 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.util.ArrayList;
 import java.util.Collection;
+import java.util.Optional;
 
 @Service
 public class CardServiceImpl implements CardService {
     private UnlockCardRequestRepository unlockCardRequestRepository;
     private CardRepository cardRepository;
+    private AccountRepository accountRepository;
     private Logger logger = LoggerFactory.getLogger(CardServiceImpl.class);
 
     @Autowired
     public CardServiceImpl(RepositoryFactory repositoryFactory) {
         this.unlockCardRequestRepository = repositoryFactory.getUnlockCardRequestRepository();
         this.cardRepository = repositoryFactory.getCardRepository();
+        this.accountRepository = repositoryFactory.getAccountRepository();
     }
 
     @Transactional
     @Override
     public Card lockCard(long cardId) {
-        Card card = cardRepository.findById(cardId).get();
-        card.setLocked(true);
-        logger.info("Card was locked");
-        return cardRepository.saveAndFlush(card);
+        Optional<Card> cardOptional = cardRepository.findById(cardId);
+        Card card = null;
+        Card temp = null;
+        if (cardOptional.isPresent()) {
+            card = cardOptional.get();
+            card.setLocked(true);
+            logger.info("Card was locked");
+            temp = cardRepository.save(card);
+        } else {
+            logger.info("There is no such card in database");
+        }
+        return temp;
     }
 
     @Transactional
     @Override
     public Card createCard(Card card, long accountId) {
-        card.getAccount().setAccountId(accountId);
-        logger.info("Creation of new card");
-        return cardRepository.save(card);
+        Optional<Account> optionalAccount = accountRepository.findById(accountId);
+        Card temp = null;
+        if (optionalAccount.isPresent()) {
+            Account account = optionalAccount.get();
+            card.setAccount(account);
+            account.getCards().add(card);
+            temp = cardRepository.save(card);
+            logger.info("Creation of new card");
+        } else {
+            logger.info("There is no such account in database");
+        }
+        return temp;
     }
 
     @Transactional
@@ -74,7 +96,7 @@ public class CardServiceImpl implements CardService {
             cards.add(cardRepository.findById(request.getCard().getId()).get());
         }
         if (cards.size() == 0) {
-            logger.info("There is no requests to unlock card");
+            logger.info("There is no such requests to unlock card");
         } else {
             logger.info("Return all request to unlock card");
         }
@@ -84,17 +106,21 @@ public class CardServiceImpl implements CardService {
     @Transactional
     @Override
     public void unlockCard(long cardId) {
-        UnlockCardRequest request = unlockCardRequestRepository.findAll()
+        Optional<UnlockCardRequest> optionalRequest = unlockCardRequestRepository.findAll()
                 .stream()
                 .filter(unlockCardRequest -> unlockCardRequest.getCard().getId() == cardId)
-                .findFirst()
-                .orElse(null);
-        Card card = cardRepository.findById(request.getCard().getId()).get();
-        card.setLocked(false);
-        cardRepository.save(card);
-        unlockCardRequestRepository.deleteById(request.getId());
-        logger.info("Card was locked");
-
+                .findFirst();
+        if (optionalRequest.isPresent()) {
+            UnlockCardRequest request = optionalRequest.get();
+            Optional<Card> optionalCard = cardRepository.findById(cardId);
+            if (optionalCard.isPresent()) {
+                Card card = optionalCard.get();
+                card.setLocked(false);
+                cardRepository.save(card);
+                unlockCardRequestRepository.deleteById(request.getId());
+                logger.info("Card was locked");
+            }
+        }
     }
 
     @Transactional
